+++ src/voxios/baseConfig.js
@@ -1,5 +1,6 @@
 const defaultConfig = {
   throwErrorMessage: true,
+  invalidToken: []
 }
 
 export default defaultConfig